4M_dataset_dev/captioning/requirements.txt
torch
transformers
pillow
bitsandbytes
accelerate